Storage of proof construction data in Solana Devnet
ProofData struct: We created a structure that represents the proof data. This structure contains the proof itself, the public inputs and the timestamp.
ProofAccount: We created a new account structure to store the proof data.
StoreProof: We created a new instruction context to store the proof data.
store_proof function: This new function will be used to store proof data in Solana Devnet.
StateAccount update: We added the proof_count field to keep track of the total number of proofs.
initialize function update: We updated proof_count to initialize with zero.
New error type: InvalidProofData error.
New constants: We set maximum values for proof and public input sizes.

A user or system sends the proof and related public inputs resulting from FHE operations by calling the store_proof function.
The program stores this data in a new ProofAccount.
A unique account is created for each proof and its address is derived using the state.proof_count value.
The hash of the proof is calculated and logged so that it can be used later for verification.
The proof_count value in StateAccount is incremented.