Control of structures required for all functions.
We added the necessary structures for all functions (Initialize, RegisterAdvertiser, CreateAd, StoreEncryptedUserData, MatchAds).
We added short explanations (documentation) for each function.
We used the require! macro to check the validity of input data.
We added new errors to the error types (InvalidAdvertiserData, InvalidAdContent, InvalidEncryptedData).
We checked all functions for errors.