update front
The fetchEncryptedUserTraits function retrieves the user's encrypted data from the Solana program, decrypts it and re-encrypts it. In a real FHE system, this will be done in a different and more secure way.
PDAs are used for program accounts (state, ads, matchedAds).
Error handling has been added and error messages are shown to the user.
Connection.confirmTransaction is used for transaction confirmation.
Matched ads data is retrieved and parse using program.account.matchedAdsAccount.fetch.